+++ lib/testing_record/model.rb
@@ -110,7 +110,6 @@ module TestingRecord
     private
 
     def reorder_attributes_for_inspect!
-      return unless self.class.__primary_key
       return if attributes.keys.first == self.class.__primary_key
 
       pk_value = attributes.delete(self.class.__primary_key)