+++ spec/testing_record/dsl/builder/settings_spec.rb
@@ -4,10 +4,6 @@ RSpec.describe TestingRecord::DSL::Builder::Settings do
   subject(:klazz) do
     Class.new do
       extend TestingRecord::DSL::Builder::Settings
-
-      def initialize(attributes = {})
-        @attributes = attributes
-      end
     end
   end
 
@@ -40,20 +36,20 @@ RSpec.describe TestingRecord::DSL::Builder::Settings do
       end
 
       it 'cannot be configured on the model' do
-        expect { FakeModel.caching :invalid }.to raise_error(TestingRecord::Error)
+        expect { FakeModel.caching :invalid }.to raise_error(TestingRecord::Error::InvalidConfigurationError)
       end
     end
   end
 
   describe '.primary_key' do
-    let(:refined_class) do
+    subject(:refined_class) do
       Class.new(klazz) do
-        primary_key :id
+        primary_key :baz
       end
     end
 
-    it 'stores a primary key value' do
-      expect(refined_class.__primary_key).to eq(:id)
+    it 'stores the primary key on the model' do
+      expect(refined_class.__primary_key).to eq(:baz)
     end
   end
 end